requirements.txt
requests
beautifulsoup4
langchain-community
google-search-results
langchain-core
langchain-openai
python-dotenv
Scrapy
scrapy-playwright
yandex_search_api==0.1.8
lxml[html_clean]